# frontend/src/types/openapi.json
{
  "components": {
    "schemas": {
...
      "EventDetailResponse": {
        "properties": {
          "current_conclusion": {
            "title": "Current Conclusion",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "event_key": {
            "title": "Event Key",
            "type": "string"
          },
          "extra": {
            "additionalProperties": true,
            "title": "Extra",
            "type": "object"
          },
          "feedback": {
            "items": {
              "$ref": "#/components/schemas/EventFeedbackItem"
            },
            "title": "Feedback",
            "type": "array"
          },
          "independent_source_count": {
            "default": 0,
            "title": "Independent Source Count",
            "type": "integer"
          },
          "independent_verification": {
            "items": {
              "$ref": "#/components/schemas/EventEvidenceGroup"
            },
            "title": "Independent Verification",
            "type": "array"
          },
          "primary_reports": {
            "items": {
              "$ref": "#/components/schemas/EventTimelineItem"
            },
            "title": "Primary Reports",
            "type": "array"
          },
          "related_discussions": {
            "items": {
              "$ref": "#/components/schemas/EventEvidenceGroup"
            },
            "title": "Related Discussions",
            "type": "array"
          },
          "snapshots": {
            "items": {
              "$ref": "#/components/schemas/EventSnapshotItem"
            },
            "title": "Snapshots",
            "type": "array"
          },
          "source_names": {
            "items": {
              "type": "string"
            },
            "title": "Source Names",
            "type": "array"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/EventTimelineItem"
            },
            "title": "Timeline",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "why_matters": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Why Matters"
          }
        },
        "required": [
          "event_id",
          "event_key",
          "title",
          "current_conclusion"
        ],
        "title": "EventDetailResponse",
        "type": "object"
      },
      "EventEvidenceGroup": {
        "properties": {
          "content_ids": {
            "items": {
              "type": "string"
            },
            "title": "Content Ids",
            "type": "array"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "key",
          "title"
        ],
        "title": "EventEvidenceGroup",
        "type": "object"
      },
      "EventFeedbackCreate": {
        "properties": {
          "content_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Id"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "title": "EventFeedbackCreate",
        "type": "object"
      },
      "EventFeedbackItem": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "title": "EventFeedbackItem",
        "type": "object"
      },
      "EventSnapshotItem": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "what_changed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "What Changed"
          },
          "why_matters": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Why Matters"
          }
        },
        "required": [
          "version",
          "title"
        ],
        "title": "EventSnapshotItem",
        "type": "object"
      },
      "EventTimelineItem": {
        "properties": {
          "content_id": {
            "title": "Content Id",
            "type": "string"
          },
          "fetched_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fetched At"
          },
          "publish_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publish Time"
          },
          "role": {
            "default": "supporting",
            "title": "Role",
            "type": "string"
          },
          "source_name": {
            "title": "Source Name",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "content_id",
          "title",
          "source_name",
          "url"
        ],
        "title": "EventTimelineItem",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TodayHighlightEvent": {
        "properties": {
          "confidence_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Score"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "event_key": {
            "title": "Event Key",
            "type": "string"
          },
          "importance_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Importance Score"
          },
          "independent_source_count": {
            "default": 0,
            "title": "Independent Source Count",
            "type": "integer"
          },
          "primary_content_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Content Id"
          },
          "section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          },
          "source_names": {
            "items": {
              "type": "string"
            },
            "title": "Source Names",
            "type": "array"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "what_changed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "What Changed"
          },
          "why_matters": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Why Matters"
          }
        },
        "required": [
          "event_id",
          "event_key",
          "title"
        ],
        "title": "TodayHighlightEvent",
        "type": "object"
      },
      "TodayHighlightsResponse": {
        "properties": {
          "date": {
            "title": "Date",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/TodayHighlightEvent"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "date"
        ],
        "title": "TodayHighlightsResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "个人化资讯监控管理系统 API",
    "title": "Personal Information Monitor",
    "version": "1.4.3"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/atoms/events/{event_id}": {
      "get": {
        "operationId": "get_event_cluster_api_atoms_events__event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Event Cluster Api Atoms Events  Event Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Event Cluster",
        "tags": [
          "atoms"
        ]
      }
    },
...
    "/api/events/today-highlights": {
      "get": {
        "description": "Return 3-8 event cards for the PIM Digest page.\n\nEmpty ``items`` means the section should be hidden. The Timeline/资讯 page\nintentionally keeps the full content timeline and does not consume this API.",
        "operationId": "get_today_highlights_api_events_today_highlights_get",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 8,
              "maximum": 8,
              "minimum": 3,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodayHighlightsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Today Highlights",
        "tags": [
          "events"
        ]
      }
    },
    "/api/events/{event_id}": {
      "get": {
        "operationId": "get_event_detail_api_events__event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Event Detail",
        "tags": [
          "events"
        ]
      }
    },
    "/api/events/{event_id}/feedback": {
      "post": {
        "operationId": "create_event_feedback_api_events__event_id__feedback_post",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventFeedbackCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventFeedbackItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create Event Feedback",
        "tags": [
          "events"
        ]
      }
    },
...
  }
}